Create 'Deck' and 'Player' classes
Without OOP it became very hard to implement a way to distribute cards
to the players' piles.
After a deck is created all the cards are distributed to each player's
'mainPile'.
A bug was introduced in this commit which sometimes results in the
'cardsDrawn' array being undefined. A seperate bug issue will be
created to track this.

Closes #5